FIX: doc_id parameters and policy-data supply.
